rename files and format locals annotation
