+++ client/src/pages/Component/Configuration/Configuration.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 import styles from 'styles/Configuration.module.css';
 import { useGlobalStore } from '../../Store/globalStateStore';
 import { MotionCollapse } from './MotionCollapse';
-import { TypeAnnotation } from './TypeDeclaration';
+import { TypeAnnotation } from './TypeAnnotation';
 import { StackVarAnnotation } from './StackVarDeclaration';
 
 const Configuration = () => {
@@ -82,11 +82,11 @@ const Configuration = () => {
 
         <MotionCollapse isOpen={isVariableAnnotationOpen}>
           {isVariableAnnotationOpen ? (
-            <>
+            <div className={styles.indentedAnnotationArea}>
               {Object.entries(currFrame.stack_data).map(([name, memoryValue]) => (
                 <StackVarAnnotation key={name} name={name} memoryValue={memoryValue} />
               ))}
-            </>
+            </div>
           ) : null}
         </MotionCollapse>
       </div>
