+++ client/src/pages/Component/Configuration/typeAnnotation.css
